+++ packages/ui-components/src/lib/components/ButtonTab.svelte
@@ -1,9 +1,9 @@
 <script lang="ts">
-  import { Button } from 'flowbite-svelte';
+	import { Button } from 'flowbite-svelte';
 
-  export let active = false;
+	export let active = false;
 </script>
 
 <Button on:click disabled={active} {...$$props}>
-  <slot />
+	<slot />
 </Button>
