+++ packages/ui-components/src/lib/components/charts/ChartTimeFilters.test.svelte
@@ -1,0 +1,8 @@
+<script>
+	import ChartTimeFilters from './ChartTimeFilters.svelte';
+	import { writable } from 'svelte/store';
+
+	export let timeDeltaStore = writable(60 * 60 * 24 * 365); // Default to 1 year
+</script>
+
+<ChartTimeFilters bind:timeDelta={$timeDeltaStore} />
